+++ cmd/crush/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/AliHamza-Coder/crush/internal/crush"
+
+func main() {
+	crush.Run()
+}